Code/Pipeline/requirements.txt
django~=3.0.3
djangorestframework==3.11.0
requests==2.24.0
psycopg2==2.8.5

python-dateutil~=2.8.1
apscheduler~=3.6.3